# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: PhraseLog API
  version: 1.0.0
  description: |
    Internal REST contract for PhraseLog v1 — the Next.js BFF ↔ Spring Boot
    interface (ADR-010). Browsers never call this API directly; they call
    Next.js route handlers, which validate the NextAuth session or anonymous
    session state and proxy here.

    Non-public endpoints require `X-Internal-Auth`: a short-lived signed (JWS)
    token minted per request by the Next.js server, carrying either the
    authenticated `user_id` or anonymous S02 `session_token` as claims.
    See docs/auth.md and ADR-010.

    Schema source of truth: data-model.md. AI behavior: AI_PIPELINE.md. Per-screen
    flow: docs/screens/sNN.md.
servers:
  - url: https://api.phraselog.app/api/v1
    description: Production
  - url: http://localhost:8080/api/v1
    description: Local dev
...
security:
  - internalAuth: []
paths:
...
  /auth/email/verification-tokens:
    post:
      tags: [auth]
      summary: Store a magic-link verification token
      description: |
        Backs the Auth.js adapter's `createVerificationToken`. Called while the
        link is being sent, before anyone has clicked anything.

        `token` is the SHA-256 hash Auth.js computes over the raw token plus
        `AUTH_SECRET` — the raw token only ever exists in the email. The backend
        stores it verbatim and must not hash again. Token values are never
        logged.

        The outstanding-token cap is checked here too, but this is the backstop,
        not the enforcement point — see
        `/auth/email/verification-tokens/quota`. Auth.js does not cancel this
        call when the send is refused, so without a cap here every refused
        attempt would still write a row, inflating the count and locking the
        address out until expiry. Refusing here is harmless because the quota
        check already stopped the mail.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identifier, token, expires]
              properties:
                identifier:
                  type: string
                  format: email
                  description: Email address. Normalized to lowercase by the backend.
                token:
                  type: string
                  maxLength: 255
                  description: Hashed token from Auth.js. Never the raw value from the email.
                expires:
                  type: string
                  format: date-time
      responses:
        "200":
          description: Stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationToken"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/email/verification-tokens/quota:
    post:
      tags: [auth]
      summary: Ask whether another magic link may be mailed to this address
      description: |
        Called by the BFF from inside `sendVerificationRequest`, **before** the
        mail reaches SES. Refuses with 429 once the address already holds 5
        unexpired tokens.

        The ordering is the whole point. Auth.js starts the send and the token
        store concurrently and awaits them together
        (`@auth/core/lib/actions/signin/send-token.js`), so a cap applied while
        storing fires after the mail is already in flight: the SES quota the cap
        exists to protect is spent anyway, and the recipient gets a link with no
        row behind it, which reads to them as an expired link. A cap in that
        position is strictly worse than no cap.

        This decides, it does not reserve — it writes nothing beyond purging
        expired rows. The store operation keeps the same cap as a backstop, for
        the concurrency reason described there. The window between this check and the store is accepted:
        closing it would need a reservation protocol for a guard that is
        deliberately crude. It is a targeted guard against using our SES quota to
        mailbox-bomb one stranger, not a general rate limiter, and it does not
        stop an attacker who varies the address.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identifier]
              properties:
                identifier:
                  type: string
                  format: email
      responses:
        "200":
          description: There is room to send
          content:
            application/json:
              schema:
                type: object
                required: [identifier, remaining]
                properties:
                  identifier:
                    type: string
                    format: email
                  remaining:
                    type: integer
                    description: Links still allowed for this address before the cap refuses.
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimitExceeded"
  /auth/email/verification-tokens/consume:
    post:
      tags: [auth]
      summary: Fetch and delete a verification token in one transaction
      description: |
        Backs the adapter's `useVerificationToken`. Delete-and-return is atomic
        so a replayed link cannot authenticate twice — the second click finds
        nothing and lands on S03's expired-link state.

        Expiry is deliberately **not** judged here. Auth.js owns the
        `Verification` error and compares `expires` itself, so an expired row is
        returned once and then gone. Judging it on both sides invites the two to
        disagree. Each call also purges rows past their expiry so the table does
        not grow without bound.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identifier, token]
              properties:
                identifier:
                  type: string
                  format: email
                token:
                  type: string
      responses:
        "200":
          description: Token found and consumed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationToken"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: |
            No such token — expired, already used, or never issued. The three are
            indistinguishable by design; the row is gone either way. The BFF maps
            this to `null`, which is Auth.js's contract for "not found". Any other
            failure status must propagate as an error instead, or an outage would
            read to the user as an invalid link.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    internalAuth:
      type: apiKey
      in: header
      name: X-Internal-Auth
      description: |
        Short-lived signed (JWS) internal service token minted by the Next.js
        BFF, binding authenticated `user_id` or anonymous S02 `session_token`
        as claims. Signed with `INTERNAL_AUTH_SECRET` shared only between
        Next.js server and Spring Boot. See docs/auth.md and ADR-010.
...
  responses:
    Unauthorized:
      description: Missing or invalid internal auth token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    ValidationError:
      description: Request body or parameters failed validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    RateLimitExceeded:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required: [error_code, user_message, developer_hint, retryable, request_correlation_id]
      properties:
        error_code:
          type: string
          example: validation_failed
        user_message:
          type: string
          description: Korean, user-safe message. Never include secrets or raw user input.
          example: 입력값을 다시 확인해 주세요.
        developer_hint:
          type: string
          description: Likely fix for agents/developers. Must not include raw user input or secrets.
          example: Validate request body and parameters before calling the service.
        retryable:
          type: boolean
          example: false
        request_correlation_id:
          type: string
          description: Matches X-Request-Correlation-Id and structured logs.
          example: 4d6f43ef-0f4a-46c7-934f-9eec4dd0e7d7

    # ── Email magic link (S03, #19) ──
    # camelCase here follows /auth/oauth/identity (#18); see the note above the
    # /auth/email/* paths.
    VerificationToken:
      type: object
      required: [identifier, token, expires]
      properties:
        identifier:
          type: string
          format: email
          description: Lowercased email address.
          example: woojoo@phraselog.app
        token:
          type: string
          description: >-
            The hashed token, echoed back. Never the raw value from the email,
            and never written to logs.
        expires:
          type: string
          format: date-time
...